Separated lambda function object into own tag `klvm.lambda`.
Target language translator could not distinguish before what to call (in
klvm.closure->): a function or a lambda. Now

    (klvm.closure-> X)

is for calling a function X represented either by a symbol or by a variable
and

    (klvm.closure-> (klvm.lambda X))

is for calling a lambda X which is addressed by target-specific manner. See
`test/stage-2.scm` for example.